+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
 Prepend new items to make git merging easier.
+- refactor(multi-autocompleter control): deprecate extraParams API in favor of a simple apiEndpoint API
 - fix(multi-choice cell): only fetch active choices
 - fix(multi-select control): only call onChange API when the value changes (and not on first-time render)
 - feat(multi-select control): dynamic height to fill containing element